requirements.txt
anyio==4.0.0
blinker==1.4
certifi==2023.7.22
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-telegram-bot==20.5
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
systemd-python==234
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
zipp==1.0.0
